Refactored logger and lowered dependencies.
Weird enough, the big problem for the dependencies was Log::Any, because
of the old version in debian stable.

However, Log::Any should be used to increase the compatibility, not to
complicate the things.

Also, the logger itself which was used in the main module didn't look
functional (as only ->debug was implemented and was dying on ->warn, for
example), so I took the liberty to refactor the Logger and use it as logger.

All existing tests pass.

However, it's still to be seen if it's actually working as advised (as this
step was a prerequisite to do it).

See https://github.com/sludin/Protocol-ACME/issues/30